requirements.txt
absl-py==2.2.2
accelerate @ git+https://github.com/huggingface/accelerate.git@8cb3ace89485af0488d93da6c080c36319cced9e
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bert-score==0.3.13
bitsandbytes==0.45.5
blis==1.3.0
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
charset-normalizer==3.4.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.8
cloudpathlib==0.21.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
datasets==3.1.0
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
durationpy==0.10
evaluate==0.4.3
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.9.0
google-auth==2.40.1
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.19
langchain-community==0.3.24
langchain-core==0.3.60
langchain-huggingface==0.1.2
langchain-openai==0.3.6
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.42
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.12
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.79.0
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
posthog==4.0.1
preshed==3.0.9
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.13
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rouge_score==0.1.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.0.1
setuptools==80.7.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.5.1
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typer==0.15.4
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.5
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
